requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.8
Django==5.1.4
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==33.3.1
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
